requirements.txt
# Importer les bibliothèques nécessaires
pandas
matplotlib
streamlit
youtube-transcript-api
google-api-python-client
isodate
python-dotenv
textblob
nltk
spacy
vaderSentiment
langchain
openai
requests
langchain_groq
csv






# pour tout installer faire dans le terminal
# pip install -r requirements.txt